// src/screens/__generated__/CommitsListPageHistoryFragment.graphql.ts
export type CommitsListPageHistoryFragment$data = {
  readonly history: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly abbreviatedOid: string;
        readonly author: {
          readonly avatarUrl: any;
          readonly name: string | null | undefined;
          readonly user: {
            readonly avatarUrl: any;
            readonly login: string;
            readonly name: string | null | undefined;
          } | null | undefined;
        } | null | undefined;
        readonly committedDate: any;
        readonly messageHeadline: string;
        readonly oid: any;
        readonly url: any;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
  };
  readonly id: string;
  readonly " $fragmentType": "CommitsListPageHistoryFragment";
};
export type CommitsListPageHistoryFragment$key = {
  readonly " $data"?: CommitsListPageHistoryFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"CommitsListPageHistoryFragment">;
};